Re #2407 Move Project out of Stack.Types.Config
