+++ package.yaml
@@ -274,6 +274,7 @@ library:
   - Stack.Types.Config.Build
   - Stack.Types.Config.Exception
   - Stack.Types.ConfigMonoid
+  - Stack.Types.Curator
   - Stack.Types.Docker
   - Stack.Types.DockerEntrypoint
   - Stack.Types.DownloadInfo
@@ -294,6 +295,8 @@ library:
   - Stack.Types.Package
   - Stack.Types.PackageName
   - Stack.Types.Platform
+  - Stack.Types.Project
+  - Stack.Types.ProjectConfig
   - Stack.Types.PvpBounds
   - Stack.Types.Resolver
   - Stack.Types.Runner
