+++ src/test/Stack/ConfigSpec.hs
@@ -27,9 +27,14 @@ import           Stack.Prelude
 import           Stack.Runners
 import           Stack.Types.BuildConfig ( BuildConfig (..), projectRootL )
 import           Stack.Types.Config
+                   ( BenchmarkOpts (..), BuildOpts (..), CabalVerbosity (..)
+                   , Config (..), ProjectAndConfigMonoid (..), TestOpts (..)
+                   , parseProjectAndConfigMonoid
+                   )
 import           Stack.Types.ConfigMonoid
                    ( ConfigMonoid (..), parseConfigMonoid )
 import           Stack.Types.GlobalOpts ( GlobalOpts (..) )
+import           Stack.Types.Project ( Project (..) )
 import           Stack.Options.GlobalParser ( globalOptsFromMonoid )
 import           System.Directory
 import           System.Environment
